requirements.txt
torch==2.5.1
typer==0.15.1
invoke==2.2.0
fastapi==0.115.6
uvicorn==0.34.0
pillow==11.0.0
torchvision==0.20.1
dvc-gdrive==3.0.1
dvc==3.58.0
dvc-gs==3.0.1
matplotlib==3.10.0
pandas==1.5.3
numpy==1.26.4
wandb==0.19.2
grpcio==1.69.0 
markdown==3.7 
tensorboard==2.18.0 
tensorboard-data-server==0.7.2 
torch_tb_profiler==0.4.3 
werkzeug==3.1.3
google-cloud-storage==2.11.0
